requirements.txt
google-api-python-client>=2.160,<3
google-auth>=2.38,<3
google-auth-oauthlib>=1.2,<2
pandas>=2.2,<3
yfinance>=0.2.54,<0.3
